Cables ready for live test

+++ src/main/java/se/sst_55t/betterthanelectricity/block/IConsumer.java
@@ -9,6 +9,6 @@ public interface IConsumer {
 
     TileEntity getOutputTE();
 
-    int getConsumeRate();
+    float getConsumeRate();
 
 }
