+++ src/main/java/se/sst_55t/betterthanelectricity/block/IGenerator.java
@@ -9,5 +9,5 @@ public interface IGenerator {
 
     TileEntity getInputTE();
 
-    int getChargeRate();
+    float getChargeRate();
 }
